requirements.txt
requests
pandas
sqlalchemy
scikit-learn
python-dotenv